repeated reordering works
This commit changes to-do list representation to a map from unique ids
to (idx,item) pairs plus a vector from idx to id (previously the to-do
list implementation was just a vector of items). I made this change
because I wanted to store the item values keyed by a unique id (which
I thought would make the drag/drop code simpler) but I then wanted an
additional data structure to improve time complexity: the drag code
involves lookup by idx (to update displaced items), so I also needed
the items to be keyed by index. I thought that this scheme would make
deletion a constant time operation, but failed to recognize that
deleting one element involves editing the entire list (decrementing
all subsequent indices). If I revert to a simple map, the code will
become simpler and shorter (and there will be a single source of
truth).

This commit also makes some small improvements like moving unique id
generation out of views, replacing the obscure :drag-prev db value to
a :selected-item value, which the item view can then use to improve
rendering of the items (both ensuring that the selected item renders
above the other items, and rendering the height of the selected item
differently than all of the other heights). There is an argument to be
made for rendering all items the same way--which involves setting
height values in the event handling code, as is currently
done--because this will allow for smooth translation of the other
items as they are displaced by the selected items. In general, if a
height value was only needed for the selected item, I would rather
store that height with :selected-item than in the items list.