reordering animation proof of concept
This commit uses react-flip-move to animate to-do list items. I relied
heavily on this project in figuring this out:

http://upgradingdave.com/blog/posts/2016-12-17-permutation.html

Note: I also relied on the following library for implementing
drag-and-drop

https://github.com/borkdude/draggable-button-in-reagent

One problem with combining drag-and-drop with react-flip-move is that
dragging-and-dropping requires you to specifically edit element
position based on cursor position, but react-flip-move tries to
abstract this away. Therefore, to figure out when to displace other
items with a dragged item, we need to inspect the react-flip-move
children to determine rendering attributes that are supposed to be
opaque. I have determined some of these values and hardcoded them into
this commit as re-frame-todo-list.views/initial and .../multiple. They
work on my current install of chrome on my machine, but are doubtless
very fragile.

A more robust method for ascertaining these values would be to track
all heights of the rendered items (doing this efficiently would
probably involve storing heights in the DB, e.g. on item
add/update, instead of populating them with every component
update). Then we could use these item heights to maintain a mapping
between item index and height to determine when the dragged item
should displace another item. Right now, we just assume that all list
items are the same height, and use a simple mx+b formula (initial +
index*multiple) to determine when to displace.

A second problem is that we need to control when and how
react-flip-move animates some items, which is not an abstraction it
provides. Currently, we are just changing the visibility of the
dragged item to "hidden" in the flip-move list, and then reanimating
it according to the cursor position. This means we have to emulate the
html/css formatting to match the flipmove react element. Instead,
maybe we could overwrite the position of the selected list item (which
gets rendered by flipmove).

Even if we manage to reposition the dragged list item (instead of
hiding it in the list and separately reanimating it), react-flip-move
only animates edits to the list, not the snapping of an off-target
dropped item to its proper position.

All in all, react-flip-move kind of works, but I'm going to explore
alternatives.

One idea is using CSS animations or keyframes or something. The
problem with keyframes is that they exist as a custom CSS class, which
spreads the view abstraction from CLJS into CSS files. This SO post
gives an example of using keyframes to animate a list in re-frame:

https://stackoverflow.com/questions/73396211/animation-of-adding-and-removing-elements-to-from-a-list-in-re-frame

But supposedly it's possible to keep CSS classes in CLJS. Aparently
there are at least 3 extensions of hiccup: sablono, hicada, and
radiant. This is an article on hicada:

https://medium.com/@rauh/a-new-hiccup-compiler-for-clojurescript-8a7b63dc5128

And here are links to the projects:

https://github.com/r0man/sablono
https://github.com/rauhs/hicada
https://github.com/BrianChevalier/radiant

The pros for animating a reorderable list using keyframes are a
high degree of customization and the fun of fully implementing the
widget in CLJS (and the fun of doing so).

Another idea is porting an existing react library to CLJS, such
as this one:

https://contactmentor.com/react-drag-drop-list/

Instructions for porting a react library to CLJS are here:

http://upgradingdave.com/blog/posts/2017-01-08-cljsjs.html

The pros of this would be learning more about react and the
satisfaction of invoking an existing solution instead of reinventing
the wheel.