suppress animation after dropping item
The item-view animation should continue to hide the dragged item after
it is dropped, until the other items complete their animation into new
places. We had to add an atom since the db only stores the selected
item--as soon as it is dropped, no item is selected and the items-view
must know which item to continue to hide.